Remove File run_init.json
UUID on 64bit

FPS Status

+++ Tomato/Tomato/src/Tomato/Core/App/UUID.h
@@ -7,13 +7,13 @@ namespace Tomato
 	{
 	public:
 		UUID();
-		explicit UUID(uint32_t uuid);
+		explicit UUID(uint64_t uuid);
 
-		uint32_t Get() const { return m_Identifier; }
+		uint64_t Get() const { return m_Identifier; }
 
 		bool operator==(const UUID& other) const;
-		operator uint32_t() const { return m_Identifier; }
+		operator uint64_t() const { return m_Identifier; }
 	private:
-		uint32_t m_Identifier;
+		uint64_t m_Identifier;
 	};
 }
